requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.13.1
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
requests==2.29.0
urllib3==1.26.15
